Fix bug w/ using jobs in zsh inside variable assignment
Zsh 5.9 appears to have introduced a bug where the first job is
not returned by `jobs` when running in a variable substitution
expression, possibly as part of a feature to report on parent
jobs from subshells.

At any rate, the jobstates variable exposes what we want here, so
no need for shell magic piping through wc and tr.