Nuke bundles from orbit and rebuild
Obviously my Vim bundle setup was a mess. Like, an awful horrific mess.
That one closet that you keep a bunch of crap you never use in, but it's
also where you keep your big suitcase, and you need to get it out, and
you dread opening the door, and when you do a bunch of stuff has shifted
and starts to collaps onto you like the stuff _knows_ that you did this
to it and it wants revenge. And you keep thinking to yourself that some
day you are gonna clean that closet. Some day soon. It can't be that bad
right? Most of the stuff in there is junk you never use, you just need
to get some trash bags and haul it all out to the curb. How long could
it take? An hour, maybe two? But you never do clean it. You just leave
it, letting it fester. As a kid you once thought there were monsters
living in the closet. As an adult, you are wiser. You _know_ there are.

But rejoyce, reader, for that day has come. The closet is clean. The
junk has been junked. And what remains is a better, simpler, more
organized world. The future is bright. You just need to keep it clean,
now, right?

...There's no way this could happen again.