Hacky fix to zsh prompt slowness
My zsh prompt was running extraordinarily slow on tymora (~650ms command
lag according to https://github.com/romkatv/zsh-bench). Profile
unsurprisingly showed most of this time was spent in vcs info functions
but surprisingly showed that the plurality of it was in hg_info. Indeed,
on tymora `hg id` is taking 0.24s just to return an error when not in an
hg repo. This also explains why this system specifically seemed extra
slow; hg is not installed on the other systems I'm regularly running zsh
on.

I don't work on any projects that use Mercurial at the moment, so I'm
removing it from the prompt as a hacky temporary fix to at least
slightly improve prompt speed on tymora. It's still noticeably slow in
git repos (~400ms total), but substantially faster when not in a repo
(<15ms).

Long term I want to fix general profile slowness here. nvm is an issue,
but so is vcs_info. It may be worth looking at powerlevel10k as an
alternative here for the custom gitstatus plugin.